@@ -31,7 +31,7 @@ classifiers = [  #! Update me
 ]
 
 [tool.poetry.dependencies]
-python = ">3.7.1,<3.11"
+python = ">=3.8.0,<4.0.0"
 importlib_metadata = {version = "^4.5.0", python = "<3.8"}
 python-dotenv = "^0.21.0"
 pydantic = ">=1.9.2,<3.0.0"
@@ -67,6 +67,9 @@ mkdocstrings = ">=0.19.1,<0.23.0"
 mkdocstrings-python = ">=0.8.3,<0.11.0"
 pillow = "^9.4.0"
 cairosvg = "^2.6.0"
+bump-pydantic = "^0.6.1"
+types-requests = "^2.31.0.2"
+vcrpy = "^5.1.0"
 
 [tool.black]
 # https://github.com/psf/black
